Streamline some projects as readme only
Include a separate list of GitHub repos in _config.yaml under "readmes". Rather than doing a full clone as we for "projects" these instead just have the README.md file fetched from GitHub.

There are a number of projects where effectively the projects page is like a paper page, which it's really just an abstract + link to the actual thing. In this case, a GitHub repo, rather than a PDF.