Add fediverse support to member layout
Because Font Awesome doesn't currently have a "fediverse" icon
available, and even the "mastodon" one wasn't added until a newer
version of the icon set than this site is using, I somewhat
arbitrarily picked the handshake icon for this section. I considered
just not having an icon, but I reviewed this section in all the
existing "team" pages and everything else has an icon, so I figured
the handsahke was better than blankness.