Update Reveal.js to v5.2.0
This uses the same pattern as previously of copying Reveal.js CSS and JS to the repo and referencing these files from a custom layout. I've updated this to talk_reveal5.html to not interfere with previous talks in case there's any incompatibilities. I also added a new citation strategy to keep the citation box fixed to the lower right of the viewport.