Upgrade from Font Awesome v4.7.0 to v.6.7.2
This is an attempt at a 1:1 refactor, though there were a few icons present in v4 that are no longer present in v6. This resulted in a handful of replacements. The goal here however is to have the greater swath of icons available and more flexible styling options from v6.